chore: update readme, code refactor, implement getMessageStatus()
