+++ contracts/src/challenge/IChallenge.sol
@@ -38,7 +38,6 @@ interface IChallenge {
     enum CompletionReason {
         OSP_VERIFIED, // OSP verified by winner.
         TIMEOUT // Loser timed out before completing their round.
-
     }
 
     event Completed(address winner, address loser, CompletionReason reason);
