+++ packages/sdk/README.md
@@ -1,1 +1,17 @@
-## The @specularl2/sdk package provides a set of tools for interacting with Specular.
+# Specular SDK
+
+## What is Specular SDK?
+
+Software Development Kit for interacting with the Specular.
+
+### What is the SDK for?
+
+The SDK helps you use the following functionalities and provides a seamless experience for getting involved with Specular as a developer.
+
+- Deposit native ETH and ERC20 tokens from an L1 network (initially Sepolia) to Specular Testnet
+- Withdraw native ETH and ERC20 tokens from Specular Testnet to an L1 network(initially Sepolia)
+- Check Specular Testnet balances
+
+You can clone the repo and play around with the example scripts.
+
+More about [Specular](https://docs.specular.network/).
