+++ packages/sdk/src/index.ts
@@ -1,3 +1,2 @@
 export * from './interfaces'
-export * from './cross-chain-messenger'
-export * from './l2-provider'
+export * from './service_bridge'
