+++ packages/sdk/src/interfaces/index.ts
@@ -1,3 +1,1 @@
-export * from './bridge-adapter'
-export * from './l2-provider'
 export * from './types'
