--- packages/sdk/src/utils/assert.ts
@@ -1,5 +1,0 @@
-export const assert = (condition: boolean, message: string): void => {
-  if (!condition) {
-    throw new Error(message)
-  }
-}
