+++ packages/sdk/src/utils/index.ts
@@ -1,3 +1,1 @@
-export * from './assert'
-export * from './coericon'
 export * from './helpers'
