+++ sbin/entrypoint.sh
@@ -9,12 +9,11 @@ cd /specular/workspace
 #     rm .*.lock
 # fi
 
-
 echo "Setting environment variables"
-export INFURA_KEY=`cat infura_pk.txt`
-export DEPLOYER_PRIVATE_KEY=`cat deployer_pk.txt`
-export SEQUENCER_PRIVATE_KEY=`cat sequencer_pk.txt`
-export VALIDATOR_PRIV_KEY=`cat validator_pk.txt`
+export INFURA_KEY=$(cat infura_pk.txt)
+export DEPLOYER_PRIVATE_KEY=$(cat deployer_pk.txt)
+export SEQUENCER_PRIVATE_KEY=$(cat sequencer_pk.txt)
+export VALIDATOR_PRIV_KEY=$(cat validator_pk.txt)
 set -o allexport
 . .sp_geth.env
 . .sp_magi.env
@@ -26,18 +25,17 @@ set +o allexport
 
 case "$1" in
 deploy)
-    # Run the main container command.
-    echo "Running deploy for genesis and JWT"
-    /specular/sbin/generate_jwt_secret.sh
-    /specular/sbin/deploy_l1_contracts.sh -y
-    ;;
+  # Run the main container command.
+  echo "Running deploy for genesis and JWT"
+  /specular/sbin/generate_jwt_secret.sh
+  /specular/sbin/deploy_l1_contracts.sh -y
+  ;;
 start)
-    shift
-    /specular/sbin/$@
-    ;;
+  shift
+  /specular/sbin/$@
+  ;;
 *)
   echo "Unknown Command"
   exit 1
   ;;
 esac
-
